fix(NoteList): react 'key' prop equal between siblings
The old way of creating a fake note array for development was using
'Array.fill()' and passing the note title as react 'key' prop.
This, however, created an conflict, since two different notes could
have the same title.

The solution was to abstract the process into a utility function,
that given a numerical value, generates a list of fake notes for
development/testing purposes.

This implementation may be changed in the future, if it turns out to
be inefficient for testing libraries.